// src/interfaces/Candidate.interface.tsx
// TODO: Create an interface for the Candidate objects returned by the API
export default interface Candidate {
     name: string | null;
     login: string | null;
     location: string | null;
     avatar_url: string | null;
     email: string | null;
     html_url: string | null;
     company: string | null;
  }
   